feat(autopay): expose protected API endpoint for manual and scheduled runs
